Update tufte.css
Changed font sizing to rems, and added media queries so that fonts scale as browser is resized in width. Also added top and bottom margins and rationalized line-heights so that sidenotes and marginnotes align to the same baseline as the main body copy.  For widths < 400px, sidenotes and marginnotes are not displayed at all.