Fix the use of sidenotes
Span elements aren't self-closed tags in HTML5 (only in XHTML).
The symbol `/` is omited in HTML5, so `<span/>` is the same as `<span>`
The browser will find a way to close those tags somewhere and that's not
something we want.

To fix this we close the tags properly and to avoid empty elements we
wrap the relevant fragment of the text (see code).

By default `.sidenote-number` adds a number `before` the element, this work
only before the changes, so `:before` is replaced with `:after` to have
the desired result.