Provides a `booktabs` class for styling tables.
The booktabs package <http://ctan.org/pkg/booktabs> provides nice rules
and spacing for tables. Tables of this style are frequently used in
academic journals.

A heavy rule is used at the top and bottom of the table.
A medium-weight rule is separates the head row from the table contents.