Reduce font size and provide new fullwidth class for code.
Wrapping with pure CSS looks like a no go [1].  It is possible to do with
Javascript, but it gets complex to do right; for example, rewrapping long lines
on browser resize.

So, take solution from @Norgat in edwardtufte/tufte-css#80 and bump the font
down to 1rem.  This gives you:

    100 characters @ body width of 1040px
     80 @ 828px
     73 @ 750px (css break)

Seems like the best we can do.

Note that stackoverflow takes the same approach with code snippets; there is no
wrapping and you have to scroll horizontally.  So this UX is what most coders
are used to on the web.

It's kind of funny that neither Microsoft's [2], Google's [3] or iOS [4] style
guides mention showing code as text!

[1] http://stackoverflow.com/a/3750513
[2] https://msdn.microsoft.com/en-us/windows/uwp/controls-and-patterns/fonts
[3] https://material.google.com/style/typography.html#
[4] https://developer.apple.com/library/ios/documentation/UserExperience/Conceptual/MobileHIG/ColorImagesText.html#//apple_ref/doc/uid/TP40006556-CH58-SW1