Make 'h2'/'h3' margin consistent with 'p' margin
The margin-top and margin-bottom on a `<p>` tag is 1.4rem.  When two
`<p>` tags are adjacent to each other, the margin bottom of the first
and the margin top of the second collapse together—they're only
separated by 1.4rem total.

On the other hand, the margin-bottom on `<h2>` and `<h3>` is 0.
Normally this is fine, because of the time a `<p>` follows the heading,
so the margin-top on the paragraph picks up the correct vertical
spacing.

The problem is when a heading isn't followed by a paragraph (let's say:
a generic <div>). If that div doesn't set margin-top like paragraphs do,
then both the heading and the div will be 0px apart. (In particular, if
there are descenders in the heading, they'll overlap!)

There are then two solutions:

1. Set a margin-bottom on the heading
2. Set a margin-top on the div

Setting a margin-top on the div essentially adds a *special case* for
the div. On the other hand, setting a margin-bottom solves the *general
case* of "vertical spacing after a heading."