`font-display: swap` for FOPST instead of FOUT/FOIT
Despite lack of universal browser support, this CSS feature seems like
a useful way to lessen the effects of waiting for our custom font to
load. This avoids FOIT (flash of invisible text) and FOUT (flash of
unstyled text) by telling the browser to use the next available font
in our `font-family` stack until our preferred font is
downloaded. This should replace FOUT with what we might call
FOPST (flash of partially styled text).

I'm also leaving open the possibility of switching to `optional`, so
that users on slow connections are not forced to download our
preferred font. Further experimentation is necessary.

Fixes #148